Archive department
There will be a checkbox in department edit window to mark department as archived.
Once department is marked as archived it will not be shown in user account directly.
It still will be able to see archived departmnets under  button action.
Frontpage widgets also will not have this department there.